add some comments
Reminder: the URL of /special.html?special_mail=celeb is just an example. In Miva the "special.html" page needs to be created first, then the script can be copied and pasted into the page's "content" section. If you're not using the "special.html" page or the "?special_mail=celeb" query-string, make sure you update that inside the script too or else it will not work.